refactor(ui): drop the St prefix from the login screen's own files
The `St` marker belongs to subturtle-ui's exported components, not to app-local
files that merely consume them.

  layouts/st-auth.vue              -> layouts/auth-redesign.vue
  components/common/StInlineNotice -> components/common/InlineNotice

`auth-redesign` is transitional: it should take `auth.vue`'s name once the last
pilotui auth screen is migrated.
