+++ frontend/components/common/InlineNotice.vue
